Add Github Discovery plugin
This adds Github Discovery which automatically pulls in repo
information. This commit also includes a processor which will add
ownership relations based on the Github roles of teams and users.

Note this will be in addition to current ownership definitions defined
in the repo catalog files, however we are considering moving entirely to
using Github to inform ownership.

Ref: ZFE-00000